let Libvirt automatically select best CPU model
The CPU model and vendor of a VM was hard-coded in cloubed. This prevented
cloubed from booting VM on hosts with different CPU. Now, cloubed specifies
in the XML representation of a Domain (aka. VM) that Libvirt should boot the VM
with a CPU identical to the host and enable all its features (eg. vmx).

Fixes #2.