Avoid a regression in domainEventRegister
There should be a regression in method domainEventRegister() in the
Python binding of Libvirt between versions 0.9.12 and 1.0.2. The
programmed callback function is never called whenever an event actually
happens.

The documentation of Libvirt API specifies that the usage of this
function is not recommended anymore and the alternative
domainEventRegisterAny() should be prefered instead.

Fixes #3.