new class VirtController for all libvirt calls
All calls to libvirt library are now done through the new VirtController
class. There are still few exceptions in StoragePool class though, but
it will be done in next commits. Python libvirt binding is only imported
in lib/VirtController.py however, it will greatly simplify mocking of
Libvirt in unit tests. It is a simple drop/replace change with matching
names for now but it may be improved but more advanced methods will be
added as soon as it could simplify the code in other parts of Cloubed.

Also remove useless Domain.snapshot() method temporarily. Should
reappear one day when properly designed (with #9).

Fixes #22.