introduction of virts/9p support
As of now, the mapped access mode is hard-coded. Since Libvirt does not
launch the KVM/QEMU virtual machines with root user (but with a
dedicated user), even with qemu:///system hypervisor, the passthrough
and squash access modes could not be used with write access to the 9p
shares. So, at least for the moment, only the mapped access mode is
supported by Cloubed.

Fixes #23.