detect conflicts on network creation
Before a network is created in Libvirt, cloubed now checks is there is
another active network with conflicting IP settings. If yes, a
CloubedException is raised in the library and the script prints the
error nicely along with the name of this boring network.

IP networks comparaison is done with the help of the netaddr library.
Therefore, a dependency has been added to this external library.

Fixes #30.